lagermanager/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
pymssql==2.3.1
python-decouple==3.8
